Rename s/notebook/notebooks
