+++ src/main/java/bio/terra/cli/command/Notebooks.java
@@ -1,6 +1,6 @@
 package bio.terra.cli.command;
 
-import bio.terra.cli.command.notebook.Create;
+import bio.terra.cli.command.notebooks.Create;
 import picocli.CommandLine;
 
 /**
@@ -8,7 +8,7 @@ import picocli.CommandLine;
  * itself; it is just a grouping keyword for it sub-commands.
  */
 @CommandLine.Command(
-    name = "notebook",
+    name = "notebooks",
     description = "Commands related to AI Notebooks in the Terra workspace context.",
     subcommands = {Create.class})
-public class Notebook {}
+public class Notebooks {}
